+++ packages/hedera-wallet-snap/packages/snap/src/facades/allowance/ApproveAllowanceFacade.ts
@@ -191,6 +191,9 @@ export class ApproveAllowanceFacade {
     } catch (error: any) {
       const errMessage = `Error while trying to approve an allowance`;
       console.error('Error occurred: %s', errMessage, String(error));
+      await SnapUtils.snapNotification(
+        `Error occurred: ${errMessage} - ${String(error)}`,
+      );
       throw rpcErrors.transactionRejected(errMessage);
     }