+++ packages/hedera-wallet-snap/packages/snap/src/facades/hts/BurnTokenFacade.ts
@@ -162,6 +162,9 @@ export class BurnTokenFacade {
     } catch (error: any) {
       const errMessage = `Error while trying to burn tokens`;
       console.error('Error occurred: %s', errMessage, String(error));
+      await SnapUtils.snapNotification(
+        `Error occurred: ${errMessage} - ${String(error)}`,
+      );
       throw rpcErrors.transactionRejected(errMessage);
     }