+++ packages/hedera-wallet-snap/packages/snap/src/facades/hts/DissociateTokensFacade.ts
@@ -161,6 +161,9 @@ export class DissociateTokensFacade {
     } catch (error: any) {
       const errMessage = `Error while trying to dissociate tokens from the account`;
       console.error('Error occurred: %s', errMessage, String(error));
+      await SnapUtils.snapNotification(
+        `Error occurred: ${errMessage} - ${String(error)}`,
+      );
       throw rpcErrors.transactionRejected(errMessage);
     }