+++ packages/hedera-wallet-snap/packages/snap/src/facades/hts/MintTokenFacade.ts
@@ -165,6 +165,9 @@ export class MintTokenFacade {
     } catch (error: any) {
       const errMessage = `Error while trying to mint tokens`;
       console.error('Error occurred: %s', errMessage, String(error));
+      await SnapUtils.snapNotification(
+        `Error occurred: ${errMessage} - ${String(error)}`,
+      );
       throw rpcErrors.transactionRejected(errMessage);
     }