+++ packages/hedera-wallet-snap/packages/snap/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [0.4.1](https://github.com/hashgraph/hedera-metamask-snaps/compare/v0.4.1...v0.4.1) (2024-05-29)
+
+### :page_with_curl: Documentation
+
+- Refer to the [Hedera Wallet Snap Wiki](https://docs.tuum.tech/hedera-wallet-snap/) for more info on how the snap works and how to integrate it into your own application.
+
+### :rocket: Features
+
+- [Added a new API to create a smart contract](https://github.com/hashgraph/hedera-metamask-snaps/issues/389)
+- [Added a new API to update a smart contract](https://github.com/hashgraph/hedera-metamask-snaps/issues/390)
+- [Added a new API to delete a smart contract](https://github.com/hashgraph/hedera-metamask-snaps/issues/391)
+- [Added a new API to call a smart contract function](https://github.com/hashgraph/hedera-metamask-snaps/issues/392)
+- [Added 3 new APIs to get smart contract function, bytecode and info](https://github.com/hashgraph/hedera-metamask-snaps/issues/394)
+- [Added a new API to execute ethereum transactions on the Hedera network](https://github.com/hashgraph/hedera-metamask-snaps/issues/393)
+
+### :bug: Bug Fixes
+
 ## [0.3.2](https://github.com/hashgraph/hedera-metamask-snaps/compare/v0.3.2...v0.3.2) (2024-04-29)
 
 ### :page_with_curl: Documentation